The special split function is needed in multiple places
